Attempt to separate out content alterations for better sharing with Preview and Runner
