+++ app/javascript/packs/runner_application.js
@@ -4,6 +4,8 @@
 // that code so it'll be compiled.
 
 require("@rails/ujs").start()
+require("../src/runner/index")
+// window.analytics = require("../src/runner_analytics") Not required for preview in Editor (exists in Runner version)
 
 const accessibleAutocomplete = require("accessible-autocomplete")
 import 'accessible-autocomplete/dist/accessible-autocomplete.min.css' 
@@ -47,7 +49,6 @@ if(autocompleteComponent) {
   });
 }
 
-//window.analytics = require("../src/analytics")
 
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
