+++ app/javascript/styles/_index.scss
@@ -1054,19 +1054,6 @@ html {
 }
 
 
-/* Emulate GovSpeak styles
- * -------------------------------- */
-.call-to-action {
-  background-color: govuk-colour("light-grey");
-  margin: 2em 0;
-  padding: 2em
-}
-
-.call-to-action > p {
-  margin: 0;
-}
-
-
 /* Override or enhance GovUK styles
  * -------------------------------- */
 .govuk-form-group--error,
