+++ app/javascript/styles/application.scss
@@ -22,4 +22,5 @@
 @import "./component_expander";
 @import "./component_warning_text";
 @import "./component_status_message";
+@import "./content";
 @import "./index";
